pyproject.toml
[tool.poetry]
name = "challenge_deep_learning"
version = "0.1.0"
description = ""
authors = ["lhzuin <luishenriquezuinruiz@gmail.com>"]
readme = "README.md"
package-mode = false



[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "=3.11.11"
hydra-core = "^1.3.2"
matplotlib = "^3.10.1"
tqdm = "^4.67.1"
pandas = "^2.2.3"
pre-commit = "^4.2.0"
wandb = "^0.19.10"
torch = "^2.7.0"
torchvision = "^0.22.0"
open-clip-torch = "^2.32.0"
transformers = "^4.51.3"
sentencepiece = "^0.2.0"
tokenizers = "^0.21.1"
accelerate = "^1.6.0"
safetensors = "^0.5.3"
bitsandbytes = "^0.45.5"
